Partially fixes Ticket #10
KMLWriter now creates a KML file with different colors for every geometry that comes from a different spatial variable (max colors are 4 but we can add more).
Also, geometries that belong to the same result are highlighted when the mouse is over one of them.

Unfortunately, that doesn't work with google maps API and every geometry has the default (blue) color.
Unfortunately, that doesn't work with geoxml extension too (different geometries do have different colors but same-result geometries are not highlighted).

You can test it with a query like this:

SELECT ?g ?s
WHERE
{
<http://geo.linkedopendata.gr/gag/id/9186> <http://geo.linkedopendata.gr/gag/ontology/έχει_γεωμετρία> ?g.
<http://geo.linkedopendata.gr/gag/id/9185> <http://geo.linkedopendata.gr/gag/ontology/έχει_γεωμετρία> ?s.
}

Select View Result:  1) On a map, 2) On a map (localhost), 3) Download and open it with google earth.