#kmlwriter Show on map (locahost) with one geometry variable should now work. (DLR queries)
The problem comes when one geometry overlaps with another.
For example:
We have geometry variables X and Y and geometries A, B and C.
The bindings are:
X=A, Y=B
X=A, Y=C

When I go over B, A is also highlighted.  When I go over C, A is also highlighted. But what should be highlighted when I go over A?
B? C? Both?

If the answer is B or C we lose a “connection”.
If the answer is both, it is wrong, because A, B and C are not in the same binding set.

Maybe we have to change the way we show the connection among the geometries.